nteractive Turkey map using Python Turtle to display coordinates of cities.
A Python-based interactive Turkey map using the Turtle module. This project allows users to click on the map to get coordinates (X, Y) and input a Turkish city's name to display its location on the map. The program aims to make learning about Turkey’s cities and coordinates fun and interactive.